moved create_restriction and restrict_csc_array to restrictions.py
