+++ src/pygeon/__init__.py
@@ -90,7 +90,11 @@ from pygeon.numerics.innerproducts import (
     lumped_peak_mass,
 )
 from pygeon.numerics.stiffness import cell_stiff, face_stiff, ridge_stiff, peak_stiff
-from pygeon.numerics.restrictions import remove_tip_dofs
+from pygeon.numerics.restrictions import (
+    remove_tip_dofs,
+    create_restriction,
+    restrict_csc_array,
+)
 from pygeon.numerics.linear_system import LinearSystem
 from pygeon.numerics.projections import eval_at_cell_centers
 from pygeon.numerics.spanningtree import (
